+++ CHANGELOG.md
@@ -6,6 +6,12 @@ All notable changes will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased <!-- 1.4.1 -->
+
+### Fixed
+
+- Fix `manager add` not detecting workspaces when an object instead of an array is used.
+
 ## [1.4.0] 26-10-2024
 
 ### Added